refactor(federation): dissolve backend-federation; hoist FederatedCredentials to core
Outbound federation lived in two crates with a backwards-feeling edge
(oidc-provider -> backend-federation). backend-federation bundled two things of
different altitude: the credential value type (foundational) and the AWS STS
exchange mechanism. With the proxy as the only consumer of the creds it obtains
-- multistore never spends a token it didn't mint nor brokers cloud access to
third parties -- the separate "spender" crate had no future tenant.

- Move FederatedCredentials (+ apply_to + redacting Debug) into multistore core
  (crates/core/src/types.rs), beside its inbound sibling TemporaryCredentials.
- Absorb the AWS AssumeRoleWithWebIdentity mechanism + FederationError into
  oidc-provider's exchange/aws.rs (no more delegation); the StsError mapping is
  now intra-crate. oidc-provider re-exports FederatedCredentials so it stays the
  single front door for consumers (source.coop never names a federation crate).
- Drop the unused body()/endpoint() helpers (body duplicated the form_pairs
  path; endpoint was never wired in) and the now-unused url dep.
- Delete crates/backend-federation; update workspace Cargo.toml, README,
  release-please config, the wrangler example, and the smoke-test docstring.

oidc-provider now depends only on multistore core. Tests: core 74, oidc-provider
29 default / 37 with azure,gcp; no new clippy warnings; cf-workers wasm passes.
