+++ Cargo.toml
@@ -7,7 +7,6 @@ members = [
     "crates/static-config",
     "crates/sts",
     "crates/oidc-provider",
-    "crates/backend-federation",
     "examples/server",
     "examples/lambda",
     "examples/cf-workers",
@@ -19,7 +18,6 @@ default-members = [
     "crates/static-config",
     "crates/sts",
     "crates/oidc-provider",
-    "crates/backend-federation",
     "examples/server",
     "examples/lambda",
 ]
@@ -110,4 +108,3 @@ multistore-metering = { path = "crates/metering", version = "0.4.0" }
 multistore-cf-workers = { path = "crates/cf-workers", version = "0.4.0" }
 multistore-oidc-provider = { path = "crates/oidc-provider", version = "0.4.0" }
 multistore-path-mapping = { path = "crates/path-mapping", version = "0.4.0" }
-multistore-backend-federation = { path = "crates/backend-federation", version = "0.4.0" }
