--- crates/backend-federation/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "multistore-backend-federation"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-description = "Outbound credential federation (OIDC identity -> backend cloud STS) for the multistore S3 proxy gateway"
-
-[dependencies]
-multistore.workspace = true
-chrono.workspace = true
-serde.workspace = true
-quick-xml.workspace = true
-url.workspace = true
-thiserror.workspace = true
