+++ release-please-config.json
@@ -44,11 +44,6 @@
           "type": "toml",
           "path": "Cargo.toml",
           "jsonpath": "$.workspace.dependencies.multistore-path-mapping.version"
-        },
-        {
-          "type": "toml",
-          "path": "Cargo.toml",
-          "jsonpath": "$.workspace.dependencies.multistore-backend-federation.version"
         }
       ]
     }
